Added tests and fixes

+++ src/chunk_metadata.rs
@@ -1,10 +1,21 @@
-use std::{io::Result as IoResult, ops::Range, path::Path, sync::Arc};
+use std::{
+    ops::Range,
+    path::Path,
+    sync::{
+        atomic::{AtomicBool, Ordering},
+        Arc,
+    },
+};
 
-use crate::{manager::DownloadManager, ChunkId, DataChunkRef, DatasetId};
+use crate::{
+    error::DownloadManagerResult,
+    manager::{BackgroundTasks, DownloadManager},
+    ChunkId, DataChunkRef, DatasetId,
+};
 
 /// Struct representing data chunk metadata. Can be retrieved by passing the chunk ID to
 /// [`crate::DataManager::find_chunk`]
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Debug)]
 pub struct DataChunkMetadata {
     id: ChunkId,
     /// Dataset that this chunk belongs to.
@@ -15,6 +26,11 @@ pub struct DataChunkMetadata {
     // Wrapped in an `Arc` to merely avoid cloning on dropping this type.
     // It's not like this path should be changing anyway.
     path: Arc<Path>,
+    /// A reference to the manager's tasks.
+    /// Used for adding the chunk removal task handle there so the manager can keep track of it.
+    manager_tasks: BackgroundTasks,
+    /// Flag from the [`DownloadManager`] signaling whether it is live and can delete chunks.
+    can_delete: Arc<AtomicBool>,
 }
 
 impl DataChunkMetadata {
@@ -30,6 +46,7 @@ impl DataChunkMetadata {
         &self.block_range
     }
 
+    /// Creates a new [`DataChunkMetadata`] instance.
     // Note that since this is not public API we can
     // afford being stricter about the `path` datatype.
     pub(crate) fn new(
@@ -37,13 +54,22 @@ impl DataChunkMetadata {
         dataset_id: DatasetId,
         block_range: Range<u64>,
         path: Arc<Path>,
+        manager_tasks: BackgroundTasks,
+        can_delete: Arc<AtomicBool>,
     ) -> Self {
         Self {
             id,
             dataset_id,
             block_range,
             path,
+            manager_tasks,
+            can_delete,
         }
+    }
+
+    /// Method that enables deletion of a chunk when the its metadata is dropped.
+    pub(crate) fn enable_deletion(&self) {
+        self.can_delete.store(true, Ordering::Release);
     }
 
     /// Method that marks the chunk directory for deletion and deletes it afterwards.
@@ -54,7 +80,7 @@ impl DataChunkMetadata {
     /// deletion, so that the [`DownloadManager`] can continue deletion when a new instance is
     /// started.
     #[tracing::instrument(fields(path = format_args!("{}", path.as_ref().display())), err(Debug))]
-    async fn delete_chunk<P>(path: P) -> IoResult<()>
+    async fn delete_chunk<P>(path: P) -> DownloadManagerResult<()>
     where
         P: AsRef<Path>,
     {
@@ -65,7 +91,9 @@ impl DataChunkMetadata {
         new_path.push(DownloadManager::DELETE_CHUNK_MARKER);
 
         tokio::fs::rename(&path, &new_path).await?;
-        tokio::fs::remove_dir_all(new_path).await
+        tokio::fs::remove_dir_all(new_path).await?;
+
+        Ok(())
     }
 }
 
@@ -77,7 +105,13 @@ impl DataChunkRef for Arc<DataChunkMetadata> {
 
 impl Drop for DataChunkMetadata {
     fn drop(&mut self) {
-        let future = Self::delete_chunk(self.path.clone());
-        tokio::spawn(future);
+        // Only delete the chunk if the download manager is live.
+        if self.can_delete.load(Ordering::Acquire) {
+            let future = Self::delete_chunk(self.path.clone());
+            self.manager_tasks
+                .lock()
+                .unwrap()
+                .push(tokio::spawn(future));
+        }
     }
 }
